feat(benchmarks): add web tooling benchmark lane (#946)
* feat(benchmarks): add web tooling benchmark lane

Add a Goccia-only V8 Web Tooling Benchmark lane that runs every pinned upstream workload, records build/execution failures as report data, and publishes retained reports to Vercel Blob on main.

Closes #857

* fix(jsx): reject relational less-than as tag start

Tighten JSX start detection so identifier-looking relational operands must also be followed by JSX tag syntax before the transformer takes over.

Add a regression for a contextual `of` identifier before `<` so minified comparison expressions keep parsing as JavaScript.

* fix(parser): handle web-tooling parser failures

* fix(parser): handle dense tooling syntax

* fix(benchmarks): run web tooling workloads independently

* fix(benchmarks): resolve relative suite dependencies

* fix(benchmarks): resolve generated workload aliases

* fix(bytecode): support large web tooling workloads

Widen compiler and VM operands used by generated tooling bundles, preserve
catch handlers across return expressions, and complete the benchmark harness
filesystem and timeout contracts.

* ci(benchmarks): budget long web tooling workloads

Use hosted-runner measurements for CoffeeScript and PostCSS timeout overrides,
and keep matrix job names stable when carrying per-workload limits.

* fix(vm): collect under bytecode memory pressure

Keep explicit near-limit collections active while routine bytecode GC is
disabled, and check for pressure at bounded VM dispatch intervals. Apply a
calibrated managed-heap ceiling to PostCSS so the full web tooling workload
stays within hosted-runner memory.

* perf(regexp): accelerate repeated tooling matches

* fix(benchmarks): stabilize all web tooling workloads

+++ source/units/Goccia.Bytecode.Debug.pas
@@ -13,7 +13,7 @@ type
 
   TGocciaLocalInfo = record
     Name: string;
-    Slot: UInt8;
+    Slot: UInt16;
     StartPC: UInt32;
     EndPC: UInt32;
   end;
@@ -30,7 +30,7 @@ type
 
     procedure AddLineMapping(const APC: UInt32; const ALine: UInt32;
       const AColumn: UInt16);
-    procedure AddLocal(const AName: string; const ASlot: UInt8;
+    procedure AddLocal(const AName: string; const ASlot: UInt16;
       const AStartPC, AEndPC: UInt32);
 
     function GetLineForPC(const APC: UInt32): UInt32;
@@ -65,7 +65,7 @@ begin
   Inc(FLineMapCount);
 end;
 
-procedure TGocciaDebugInfo.AddLocal(const AName: string; const ASlot: UInt8;
+procedure TGocciaDebugInfo.AddLocal(const AName: string; const ASlot: UInt16;
   const AStartPC, AEndPC: UInt32);
 begin
   if FLocalCount >= Length(FLocals) then
